Fix interlace issues
We now copy the lines of the previous field into the current field, so
every frame contains both fields, but only one of the fields are new.

There are lots of hardcoded values in the somagic_video.c file in this
commit, and the driver is also mostly hardcoded to PAL.

You will need to run your capture application at 50FPS for this version
of the driver to work.

In mplayer, the videofilter pp=lb seems like a good option.

example:
> mplayer -vf pp=lb -vo x11 -fps 50 tv://norm=PAL -tv \
>	driver=v4l2:device=/dev/video1:norm=pal