Fix NTSC setup of the SAA7113 chip
Address 0x10 of the SAA7113 chip was set to 0x40 for ntsc.
That will only give us 480 lines of active video for NTSC.

The ITU-656 standard says that NTSC should be 487 lines of active
video according to the SAA7113 datasheet.

This commit makes the chip act according to the ITU-656 when parsing
NTSC video.